Fix read performance regression in ObjectCacher
The regression was introduced in commit
d7648624451fe25d72a623019c1a3badd5ab8e71.  The problem is that the cache
thinks it's full (when it's not), so it defers the read.  This change
frees up cache space if necessary and only defers the read if enough
space cannot be freed.

Fixes: 9513
Signed-off-by: Adam Crume <adamcrume@gmail.com>
(cherry picked from commit 365b007bd8fd793a9cabb16d517ef1643d951e40)