Revert "osd: detect (some) misordered ondisk tmaps"
This reverts commit 279b3425764d758ea1e2eaf563f6e007c36f8b09.

The check for 'nextkey < last_disk_key' makes not much sense since
last_disk_key is an empty string and not set before. Comparing a
decoded string to be less than an empty string will be never true.

Since this if() isn't part of a loop last_disk_key is only set
once and there is no other consumer: revert this dead code.

Conflicts:
	src/osd/ReplicatedPG.cc