mon/OSDMonitor: do not reencode OSDMap on 'osd getmap'
Return the bits unmolested.

This is a slight change in that we do not reencode based on the client's
features.  However, this is an admin command and the client is generally
the CLI or REST API.  It makes more sense to provide the unmolested
bits than to reencode them based on features present in a tool that
never inspects the map.

Signed-off-by: Sage Weil <sage@redhat.com>