osd: force read bit for watch
We can get an bad reply like this:

     - watch1 received, repop started
     - con resets
     - watch2 received
     - watch2 sees watch in oi
       - replies immediately
       - does side-effects (connects new session/con)
     - watch1 repop completes
       - replies (dropped)
       - does side-effects (connects old session/con)

Two problems: we end up with the wrong session/con connected, and the watch
reply goes out before the watch is actually stable on disk.

Fix this my marking watch ops as reads.  This will make us wait for the
previous watch addition to commit before we process the next one.  Somewhat
overkill, but sufficient.

Fixes: #10441
Signed-off-by: Sage Weil <sage@redhat.com>