osd: enable filestore_extsize by default
Note that this will only get used if the kernel is new enough; if it is
older than 3.5 the option will get disabled and extsize will not be used
even if the option is set to true.

This partially reverts ab3b5b645f29ab3b4de8ded00ca3de0c7ce38fd6.

Fixes: #9956
Signed-off-by: Sage Weil <sage@redhat.com>