mds: CDir: create scrub_info_t and surrounding scrub infrastructure
This is set up so that the coming-shortly scrub system can initialize scrubbing
on a directory and perform the actual scrubs, but rely on the directory to
keep track of its internal scrub state and selection of dentries.

Note that initialize_scrub() and scrub_local()  can only be called while the
directory is complete.

Signed-off-by: Greg Farnum <gfarnum@redhat.com>