mds: wait for mydir replicas to go away in stopping
Previously, while we were waiting for locks related
to mydir to be acknowledged by other ranks, we could
advance to the point of capping the log.  When the
locks were acked, we would try to write to the log
and assert out.  Avoid this by waiting for mydir
to not be replicated before capping the log.

Fixes: #10743
Signed-off-by: John Spray <john.spray@redhat.com>