osd/ReplicatePG: correctly checking if an object is in hit set
For the write case when the oid.snap is CEPH_NOSNAP, the obc is created
in find_object_context, and missing_oid is setting to a NULL object.
Checking if the missing_oid is in the hit set is not the correct way
since it's NULL. Instead, we should check the obc->obs.oi.soid.

Signed-off-by: Zhiqiang Wang <zhiqiang.wang@intel.com>