rbd cli: remove erroneous arg for ceph_argparse_witharg
Remove erroneous arg for ceph_argparse_witharg call when '--io-pattern' parsed:
the name look up will point the compiler to bool ceph_argparse_witharg(
    std::vector<const char*> &args,
    std::vector<const char*>::iterator &i, std::string *ret, ...) when compiler
is resolving this function call. The &err argument will be wrongly interpreted
as a char * variable to be compared with the argument name pointed by i.

Signed-off-by: Dmitry Yatsushkevich <dyatsushkevich@mirantis.com>