FileStore: sync object_map when we syncfs
This way, even if leveldb/rocksdb is on another filesystem/disk,
it will still be synced before trimming the journal.  Note, this
patch only syncs the object_map in the !backend->can_checkpoint()
case since if the backend can checkpoint, the checkpoint must
also capture the object_map since we have not been setting replay
guards in the object_map.

Signed-off-by: Samuel Just <sjust@redhat.com>