calamari_setup: handle RHCS 1.2 (missing /mnt/MON)
In RHCS 1.2 we don't have a /mnt/MON directory. The intention of
0674855b0facf1c82d2e25ca92f2d4d777d10f6a was to handle this condition,
but in 1.2, the non-zero return code makes Teuthology fail the whole
test.

We don't want *Teuthology* itself to act on the return code here; we
simply want to know what it was and structure the rest of the test
accordingly.