ceph-disk: move update_partition from main_prepare to prepare_dev
The update_partition call in main_prepare happens immediately after
prepare_dev but only if the data argument is a block device. There is no
reason for this separation: it is more sensible to call it from within
prepare_dev.

There is an additional test in prepare_dev that verifies partprobe won't
be called on a partition because it would not make sense.

Signed-off-by: Loic Dachary <ldachary@redhat.com>