mon/AuthMonitor: don't crash on ENOENT
When MonDBStore returns ENOENT when trying to get mkfs keyring, monitor
crashes with assert. On the other hand, if it returned 0 and empty
bufferlist, all was fine - so make it decode the bufferlist only when
MonDBStore returned 0, don't fail on ENOENT (and don't try to decode
empty bufferlist in that case), and fail hard otherwise.

Signed-off-by: Piotr Dałek <piotr.dalek@ts.fujitsu.com>