osdc/Objecter: end of pg if next pointer is MAX, not 0 items
It is possible in pathological cases for the OSD to return 0
items with a non-MAX next value (indicating the objecter should
ask for more).  In particular, this is easy to reproduce with
max_entries = 1 as the OSD will list and ignore the pgmeta object,
return 0 items, and have next = the next object.

I check back as far as dumpling and the OSD is setting next=MAX
at the end of the PG (and result = 1, FWIW), so using next
here instead of inferring anything from the number of result
items is safe.

Signed-off-by: Sage Weil <sage@redhat.com>