Script and Guidelines for mirroring Ceph
This commit introduces a script which admins can use to mirror
the Ceph packages to their local systems.

With this script they can easily sync from a mirror local to them.

The README explains user on how and when to sync the sources and
how they can become a official mirror for Ceph.

Signed-off-by: Wido den Hollander <wido@42on.com>