mon: remove conflicting subscribes when handle_subscribe
For monitoring log, when changing between different levels, should
remove previous level in the subscribes. This is because unlike other
PaxosServices, LogMonitor has multiple subscription names/types and
they are log levels, and hence mutual exclusive. when setting a new
subscription for logmon, we need to remove the old log level if any.

Signed-off-by: Zhiqiang Wang <wonzhq@hotmail.com>