mds: fix flushing caps check
Client may flush and drop caps at the same time. If client need to
send cap reconnect before the caps get flushed. The issued caps in
the cap reconnect does not include the flushing caps. When choosing
lock states we should consider the flushing caps.

The check for caps haven't been flushed is wrong, fix it.

Signed-off-by: Yan, Zheng <zyan@redhat.com>