osd: handle boot racing with NOUP set + clear
Resolve this race:

- osd sends boot message
- mon sets NOUP
- mon receives and drops boot message
- mon later clear NOUP
-> osd still waiting

by restarting the boot process if we are is_booting() and we see
the NOUP flag cleared.  This implies we sent the boot before NOUP
was set (we don't send boot while it is) and the mon probably
dropped our request.  If it didn't we'll go through another boot
cycle, but that is rare, mostly harmless, and unavoidable (there's
no easy way to tell whether our message was dropped or not).

Signed-off-by: Sage Weil <sage@redhat.com>