os/bluestore: change block file mkfs behavior
Previously, if path was set, we'd make a symlink.  Otherwise, if
size was set, we'd create a file and resize it accordingly. This
means that setting the size means we create the block "device"
files, which is only useful for debugging, and we want to set a
size that can be used by ceph-disk when creating partitions.

Instead, if path is set, make a symlink.  Then/also, if size is
set, and the file/symlink points to a regular file, and that
regular file is 0 bytes, then resize it.  This way, vstart.sh
(or a dev) can just touch the file and then mkfs will size it up.

Signed-off-by: Sage Weil <sage@redhat.com>