osd: consider high/low mode when putting agent to sleep
If we are in low flush mode, we may only get up to max_low_ops in flight,
in which case we may never go to sleep here.

Fix it by using the max_low_ops threshold when appropriate.

Note that agent_work() might start up *more* than this many ops (if there
are lots of evicts to do) currently, but I think it is fine if evicts go
a bit slower if we are in low mode.  (Really, the high/low shouldn't be
tied to flushing specifically.)

Fixes: #14752
Reported-by: Markus Blank-Burian <burian@muenster.de>
Signed-off-by: Sage Weil <sage@redhat.com>