tasks/scrub_test: restore the changes to omap after the test
because the repair for digest mismatch will rewrite the omap digest
in the corresponding OI attribute with the one calculated from the
omap read from disk. this is fine. but the succeeding test can not
use the same way used before to mess up the victim object in hope
to get an inconsistent object, as we have accepted it as an authroized
copy and populate the change to its replicas. to address this
problem, the crime scene of the `test_repair_bad_omap()` is cleaned
up before calling `repair()`. it helps with isolating the different
tests, so new tests should not worry about what the current omap
should look like.

Signed-off-by: Kefu Chai <kchai@redhat.com>